+++ src/test/kotlin/io/kjson/pointer/IndexOpTest.kt
@@ -1,0 +1,56 @@
+/*
+ * @(#) IndexOpTest.kt
+ *
+ * kjson-pointer  JSON Pointer for Kotlin
+ * Copyright (c) 2022 Peter Wall
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package io.kjson.pointer
+
+import kotlin.test.Test
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
+import kotlin.test.expect
+
+import io.kjson.JSON.asInt
+import io.kjson.JSONObject
+
+class IndexOpTest {
+
+    @Test fun `should use indexing operation`() {
+        val json = JSONObject.build {
+            add("abc", 123)
+            add("xyz", 789)
+        }
+        val ptr = JSONPointer("/abc")
+        expect(123) { json[ptr].asInt }
+    }
+
+    @Test fun `should use in operation`() {
+        val json = JSONObject.build {
+            add("abc", 123)
+            add("xyz", 789)
+        }
+        assertTrue(JSONPointer("/abc") in json)
+        assertFalse(JSONPointer("/def") in json)
+    }
+
+}